Move flows into packages
