+++ pkg/v3/flows/conditional/conditional_test.go
@@ -1,4 +1,4 @@
-package flows
+package conditional
 
 import (
 	"context"
@@ -11,15 +11,16 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
+	common "github.com/smartcontractkit/chainlink-common/pkg/types/automation"
+
 	ocr2keepersv3 "github.com/smartcontractkit/chainlink-automation/pkg/v3"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/service"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/stores"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/types"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/types/mocks"
-	common "github.com/smartcontractkit/chainlink-common/pkg/types/automation"
 )
 
-func TestRecoveryFinalization(t *testing.T) {
+func TestConditionalFinalization(t *testing.T) {
 	upkeepIDs := []common.UpkeepIdentifier{
 		common.UpkeepIdentifier([32]byte{1}),
 		common.UpkeepIdentifier([32]byte{2}),
@@ -77,11 +78,12 @@ func TestRecoveryFinalization(t *testing.T) {
 			WorkID:   workIDs[1],
 		},
 	}, nil).Times(times)
+
 	upkeepStateUpdater.On("SetUpkeepState", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 	// set the ticker time lower to reduce the test time
-	recFinalInterval := 50 * time.Millisecond
+	interval := 50 * time.Millisecond
 	pre := []ocr2keepersv3.PreProcessor[common.UpkeepPayload]{coord}
-	svc := newFinalRecoveryFlow(pre, rStore, runner, retryQ, recFinalInterval, proposalQ, payloadBuilder, upkeepStateUpdater, logger)
+	svc := NewFinalConditionalFlow(pre, rStore, runner, interval, proposalQ, payloadBuilder, retryQ, upkeepStateUpdater, logger)
 
 	var wg sync.WaitGroup
 	wg.Add(1)
@@ -100,19 +102,16 @@ func TestRecoveryFinalization(t *testing.T) {
 		assert.NoError(t, svc.Start(ctx))
 	}(svc, context.Background())
 
-	time.Sleep(recFinalInterval*time.Duration(times) + recFinalInterval/2)
+	time.Sleep(interval*time.Duration(times) + interval/2)
 
 	assert.NoError(t, svc.Close(), "no error expected on shut down")
 
-	coord.AssertExpectations(t)
-	runner.AssertExpectations(t)
-	payloadBuilder.AssertExpectations(t)
 	rStore.AssertExpectations(t)
 
 	wg.Wait()
 }
 
-func TestRecoveryProposal(t *testing.T) {
+func TestSamplingProposal(t *testing.T) {
 	upkeepIDs := []common.UpkeepIdentifier{
 		common.UpkeepIdentifier([32]byte{1}),
 		common.UpkeepIdentifier([32]byte{2}),
@@ -128,9 +127,13 @@ func TestRecoveryProposal(t *testing.T) {
 
 	runner := new(mocks.MockRunnable)
 	mStore := new(mocks.MockMetadataStore)
-	recoverer := new(mocks.MockRecoverableProvider)
+	upkeepProvider := new(mocks.MockConditionalUpkeepProvider)
+	ratio := new(mocks.MockRatio)
 	coord := new(mocks.MockCoordinator)
 
+	ratio.On("OfInt", mock.Anything).Return(0, nil).Times(1)
+	ratio.On("OfInt", mock.Anything).Return(1, nil).Times(1)
+
 	coord.On("PreProcess", mock.Anything, mock.Anything).Return([]common.UpkeepPayload{
 		{
 			UpkeepID: upkeepIDs[0],
@@ -143,6 +146,7 @@ func TestRecoveryProposal(t *testing.T) {
 			WorkID:   workIDs[1],
 		},
 	}, nil).Times(1)
+	coord.On("PreProcess", mock.Anything, mock.Anything).Return(nil, nil)
 
 	runner.On("CheckUpkeeps", mock.Anything, mock.Anything, mock.Anything).Return([]common.CheckResult{
 		{
@@ -158,53 +162,49 @@ func TestRecoveryProposal(t *testing.T) {
 			Eligible: true,
 		},
 	}, nil).Times(1)
+	runner.On("CheckUpkeeps", mock.Anything).Return(nil, nil)
 
 	mStore.On("ViewProposals", mock.Anything).Return([]common.CoordinatedBlockProposal{
 		{
 			UpkeepID: upkeepIDs[2],
 			WorkID:   workIDs[2],
 		},
-	}, nil).Times(2)
+	}, nil)
 	mStore.On("AddProposals", mock.Anything).Return(nil).Times(2)
 
-	recoverer.On("GetRecoveryProposals", mock.Anything).Return([]common.UpkeepPayload{
+	upkeepProvider.On("GetActiveUpkeeps", mock.Anything).Return([]common.UpkeepPayload{
 		{
 			UpkeepID: upkeepIDs[0],
 			WorkID:   workIDs[0],
 		},
-	}, nil).Times(1)
-	recoverer.On("GetRecoveryProposals", mock.Anything).Return([]common.UpkeepPayload{
 		{
 			UpkeepID: upkeepIDs[1],
 			WorkID:   workIDs[1],
 		},
-	}, nil).Times(1)
+	}, nil).Times(2)
+	upkeepProvider.On("GetActiveUpkeeps", mock.Anything).Return([]common.UpkeepPayload{}, nil)
 	// set the ticker time lower to reduce the test time
-	interval := 50 * time.Millisecond
 	pre := []ocr2keepersv3.PreProcessor[common.UpkeepPayload]{coord}
-	stateUpdater := &mockStateUpdater{}
-	svc := newRecoveryProposalFlow(pre, runner, mStore, recoverer, interval, stateUpdater, logger)
+	svc := NewSampleProposalFlow(pre, ratio, upkeepProvider, mStore, runner, time.Millisecond*100, logger)
 
 	var wg sync.WaitGroup
 	wg.Add(1)
 
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
+	defer cancel()
+
 	go func(svc service.Recoverable, ctx context.Context) {
 		defer wg.Done()
 		assert.NoError(t, svc.Start(ctx))
-	}(svc, context.Background())
-
-	time.Sleep(interval*time.Duration(2) + interval/2)
+	}(svc, ctx)
 
 	assert.NoError(t, svc.Close(), "no error expected on shut down")
 
-	mStore.AssertExpectations(t)
-	recoverer.AssertExpectations(t)
-	runner.AssertExpectations(t)
-	coord.AssertExpectations(t)
-
 	wg.Wait()
-}
 
-type mockStateUpdater struct {
-	common.UpkeepStateUpdater
+	mStore.AssertExpectations(t)
+	upkeepProvider.AssertExpectations(t)
+	coord.AssertExpectations(t)
+	runner.AssertExpectations(t)
+	// ratio.AssertExpectations(t)
 }
