+++ pkg/v3/flows/log/log_trigger_flow_test.go
@@ -1,4 +1,4 @@
-package flows
+package log
 
 import (
 	"context"
@@ -11,24 +11,35 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
+	ocr2keepersv3 "github.com/smartcontractkit/chainlink-automation/pkg/v3"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/service"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/stores"
-	"github.com/smartcontractkit/chainlink-automation/pkg/v3/types"
 	"github.com/smartcontractkit/chainlink-automation/pkg/v3/types/mocks"
 	common "github.com/smartcontractkit/chainlink-common/pkg/types/automation"
 )
 
-func TestRetryFlow(t *testing.T) {
+func TestLogTriggerFlow(t *testing.T) {
 	logger := log.New(io.Discard, "", log.LstdFlags)
 
-	times := 3
+	times := 2
 
 	runner := new(mocks.MockRunnable)
 	rStore := new(mocks.MockResultStore)
 	coord := new(mocks.MockCoordinator)
-	upkeepStateUpdater := new(mocks.MockUpkeepStateUpdater)
 	retryQ := stores.NewRetryQueue(logger)
+	upkeepStateUpdater := new(mocks.MockUpkeepStateUpdater)
+	lp := new(mocks.MockLogEventProvider)
 
+	lp.On("GetLatestPayloads", mock.Anything).Return([]common.UpkeepPayload{
+		{
+			UpkeepID: common.UpkeepIdentifier([32]byte{1}),
+			WorkID:   "0x1",
+		},
+		{
+			UpkeepID: common.UpkeepIdentifier([32]byte{2}),
+			WorkID:   "0x2",
+		},
+	}, nil).Times(times)
 	coord.On("PreProcess", mock.Anything, mock.Anything).Return([]common.UpkeepPayload{
 		{
 			UpkeepID: common.UpkeepIdentifier([32]byte{1}),
@@ -55,35 +66,24 @@ func TestRetryFlow(t *testing.T) {
 	rStore.On("Add", mock.Anything).Times(times)
 	upkeepStateUpdater.On("SetUpkeepState", mock.Anything, mock.Anything, mock.Anything).Return(nil)
 	// set the ticker time lower to reduce the test time
-	retryInterval := 50 * time.Millisecond
+	logInterval := 50 * time.Millisecond
 
-	svc := NewRetryFlow(coord, rStore, runner, retryQ, retryInterval, upkeepStateUpdater, logger)
+	svc := NewLogTriggerFlow([]ocr2keepersv3.PreProcessor[common.UpkeepPayload]{coord},
+		rStore, runner, lp, logInterval, retryQ, upkeepStateUpdater, logger)
 
 	var wg sync.WaitGroup
 	wg.Add(1)
-
-	err := retryQ.Enqueue(types.RetryRecord{
-		Payload: common.UpkeepPayload{
-			UpkeepID: common.UpkeepIdentifier([32]byte{1}),
-			WorkID:   "0x1",
-		},
-	}, types.RetryRecord{
-		Payload: common.UpkeepPayload{
-			UpkeepID: common.UpkeepIdentifier([32]byte{2}),
-			WorkID:   "0x2",
-		},
-	})
-	assert.NoError(t, err)
 
 	go func(svc service.Recoverable, ctx context.Context) {
 		assert.NoError(t, svc.Start(ctx))
 		wg.Done()
 	}(svc, context.Background())
 
-	time.Sleep(retryInterval*time.Duration(times) + retryInterval/2)
+	time.Sleep(logInterval*time.Duration(times) + logInterval/2)
 
 	assert.NoError(t, svc.Close(), "no error expected on shut down")
 
+	lp.AssertExpectations(t)
 	coord.AssertExpectations(t)
 	runner.AssertExpectations(t)
 	rStore.AssertExpectations(t)
